fix/rework nonce + remove be
